third_party/blink/web_tests/platform/generic/http/tests/devtools/load-file-resource-for-frontend-expected.txt

Test loading file resource from front-end 


Running: testResourceFromFileScheme
Loading resource from resources/file-resource-to-load-for-frontend.txt
Content: SUCCESS
